feat: fix BigInt types, add integration tests, final cleanup
- Fix u64 fields to use BigInt (consistent with solana_kit_token pattern)
- Fix MetadataArgs instruction builders to use raw bytes encoding
- Add comprehensive integration tests (12 tests)
- Fix all type mismatches and import errors
- 69 tests passing (32 unit + 14 spl_account_compression + 11 keccak256 + 12 integration)

+++ packages/solana_kit_mpl_bubblegum/lib/src/burn_asset.dart
@@ -29,7 +29,7 @@ class BurnInput {
   final List<int> creatorHash;
 
   /// The nonce (asset ID) of the compressed NFT.
-  final int nonce;
+  final BigInt nonce;
 
   /// The index of the leaf in the Merkle tree.
   final int index;
